+++ src/realm/object-store/impl/realm_coordinator.cpp
@@ -1306,14 +1306,3 @@ void RealmCoordinator::write_copy(StringData path, const char* key)
 {
     m_db->write_copy(path, key);
 }
-
-void RealmCoordinator::async_request_write_mutex(Realm& realm)
-{
-    auto tr = Realm::Internal::get_transaction_ref(realm);
-    m_db->async_request_write_mutex(tr, [realm = realm.shared_from_this()]() mutable {
-        auto& scheduler = *realm->scheduler();
-        scheduler.invoke([realm = std::move(realm)] {
-            Realm::Internal::run_writes(*realm);
-        });
-    });
-}